refactor(plugin-audit-trail): remove InMemoryAuditStore from public API
Keep it as an internal test utility only — clients have no use for it.
