+++ packages/plugin-audit-trail/src/index.ts
@@ -1,5 +1,4 @@
 export { default as auditTrail, REDACTED } from './audit-trail';
-export { default as InMemoryAuditStore } from './in-memory-store';
 export {
   createSqlAuditStore,
   ensureAuditStorage,
